Remove Rust toolchain setup from Github action for linting
The actions-rs/toolchain action respects toolchain files, and will use
them if present.

Since we currently aren't doing anything exotic with formatting in CI,
it makes sense to keep this configuration nearer to the code, rather
than having it live solely in our CI configuration. This just ensures
that things are as orthogonal and composable as possible.