[CI] Ensure Docker builds obey .cargo/config.toml settings
Previously, `.cargo` was specifically excluded from Docker
builds. This seems to be related to how we use `cargo vendor` builds.

In particular, the output of `cargo vendor` was saved to
`.cargo/config`. This file was the name of the configuration file
prior to Rust 1.39, but the current preferred name is `cargo.toml`. If
both files are present, `config` is used and `config.toml` is ignored.

Because of this, the configuration added in #544 was not being honored
in any of the container-based build systems.

Here, we stop ignoring `.cargo` in builds, allowing us to copy
`.cargo/config.toml` into the container. Then, we append the output of
`cargo vendor` to this file.

For reference, this is what that appended output looks like:

    [source.crates-io]
    replace-with = "vendored-sources"

    [source.vendored-sources]
    directory = "vendor"

The end result is the same vendoring behavior as before, but with
compilation flags honored.