Add integration feature to sessiondb_integration_tests
Previously, all the tests in this test binary were declared only for
the `integration` feature, but the helper functions were not. Since
the test code was only compiled with the `integration` feature
enabled, the helpers appeared dead otherwise.

Now, rather than tagging everything, we use an _inner_
attribute (declared with `#!` rather than `#`) to apply to the entire
file.

Signed-off-by: Christopher Maier <chris@graplsecurity.com>