[Chore] Add run-rustfmt.sh script
Similar to the `run-clippy.sh` script added in #550, this commit
introduces a `run-rustfmt.sh` script. It is intended to be the source
of truth for how we format our Rust code, regardless of the
environment.

In particular, it is designed to work on the terminal and in CI for
basic "is the code formatted?" checks, but also in editor
integrations, allowing for format-on-save actions. For example, to
configure Visual Studio Code, assuming the use of the [Rust Analyzer
Plugin][1], you would do something like the following:

    "settings": {
        "editor.formatOnSave": true,
        "editor.formatOnPaste": true,
        "[rust]": {
            "editor.defaultFormatter": "matklad.rust-analyzer",
        },
        "rust-analyzer.linkedProjects": [
            "src/rust/Cargo.toml"
        ],
        "rust-analyzer.rustfmt.overrideCommand": [
            "/path/to/your/grapl/checkout/src/rust/bin/run-rustfmt.sh"
        ],
        // ...
    }

[1]: https://marketplace.visualstudio.com/items?itemName=matklad.rust-analyzer

Signed-off-by: Christopher Maier <chris@graplsecurity.com>