Delete node identifier service test
This test was introduced in PR #485, but was always essentially
disabled and never ran. Making the test run revealed test failures,
however.

After discussion in Slack, we decided that while the ultimate goal of
the test is a good one, the implementation here is lacking. As such,
we're deleting the code now, but will revisit in [later][1]

[1]: https://github.com/grapl-security/issue-tracker/issues/218

Signed-off-by: Christopher Maier <chris@graplsecurity.com>