[Formatting] Let rustfmt keep our use statements organized
This introduces a number of related formatting directives (many of
which are currently unstable) that keep `use` statements automatically
organized.

The specific configurations are ones that I've used in the past, and
personally find to create the easiest-to-read import sections.

Since the changes touch basically every Rust file, and they
are all related, I opted to introduce them all at once, rather than
one-by-one.

The directives are:
- [reorder_imports](https://rust-lang.github.io/rustfmt/?version=master&search=#reorder_imports)
- [reorder_modules](https://rust-lang.github.io/rustfmt/?version=master&search=#reorder_modules)
- [group_imports](https://rust-lang.github.io/rustfmt/?version=master&search=#group_imports)
- [imports_indent](https://rust-lang.github.io/rustfmt/?version=master&search=#imports_indent)
- [imports_layout](https://rust-lang.github.io/rustfmt/?version=master&search=#imports_layout)
- [merge_imports](https://rust-lang.github.io/rustfmt/?version=master&search=#merge_imports)

These are all set up in the new `.rustfmt.toml` file.

Signed-off-by: Christopher Maier <chris@graplsecurity.com>