Make run-rustfmt.sh script more robust
The previous test wasn't actually what we wanted to be testing... it
only tested to see if standard input was a terminal or not, and _not_
whether standard input had any content. This ended up breaking in CI,
and not actually running our formatting checks!

To fix this, and to also help out with the usecase of having to update
_all_ Rust code at once whenever we update our configuration options,
we now add a few command-line options to determine how we want to
behave. Everything is fully documented in the script now (just pass
`--help`).

As a result, we now run with the `--check` option in CI. We also add
output on stderr to get useful feedback on what is happening.

Signed-off-by: Christopher Maier <chris@graplsecurity.com>