Reorganize protobuf packaging to be more standard and future-proof
Puts the existing protobuf types we have into the new
`graplinc.grapl.api.graph.v1beta1` package structure.

Here is the logic behind the packaging scheme:

* `graplinc`: our company; everything we ever create will live under this
namespace.

* `grapl`: the product the protobufs apply to. Right now we only have one
product, but this structure provides a natural place for anything we
may do in the future.

* `api.graph`: the sub-component of Grapl that the protobufs apply
to. There can be as many packages here as needed.

* `v1beta1`: We'll begin versioning our protobufs to provide a way to
manage any breaking changes that we may need to do, as well as provide
a place to experiment with new features.

The directory structure also reflects this package structure.

Signed-off-by: Christopher Maier <chris@graplsecurity.com>