Update Python code to use new Protobuf packaging
Additionally, our Python build logic has been modified to create a new
package based on this new structure. In order to keep flexibility, we
use namespaces to allow us to distribute subsections of the protobufs
as standalone packages, as well as provide a top-level strcuture for
the rest of our Python packages to reside in.

Signed-off-by: Christopher Maier <chris@graplsecurity.com>