Add black and isort to Pants for Python linting and formatting
We already use `black` and `isort` for linting and formatting, but now
we can let Pants handle that.

To check that all Python code conforms to our desired format, run:
```sh
./pants lint ::
```

If any violations are found, they can be fixed with:
```sh
./pants fmt ::
```

For some reason, the Pants Black plugin only seems to behave properly
if the `pyproject.toml` file is at the root of the repository; using
the original in `src/python` resulted in no files being formatted. The
`isort` plugin didn't have this problem.

In any event, placing the `pyproject.toml` configuration for these
tools at the root should help make it easier to integrate with editor configurations.

Additionally, we can use `isort`'s profiles to simplify its configuration.

(The old `pyproject.toml` file is left in place for now to facilitate
the transition to Pants; it will eventually be removed entirely.)

Signed-off-by: Christopher Maier <chris@graplsecurity.com>