Initial use of Pants to manage Python monorepo assets
[Pants][1] is a language-agnostic build platform designed to support
monorepos. It is particularly useful for Python.

It does not require code to be in any particular layout to be used,
and can thus be introduced incrementally to a codebase. This is
intended to be one in a series of commits and PRs that gradually
migrates our Python system to be driven more and more by Pants.

This initial commit introduces lots of Pants metadata files that
describe the layout and interdependencies of our code. It doesn't "do"
anything just yet, like run tests or build Python wheels. You can,
however, do even some basic introspection tasks now, like:

* Get a report of what files Pants knows about:
``` sh
./pants filedeps ::
```

* See how many lines of code are in the project (Python, at least):
``` sh
./pants count-loc ::
```

* View all 3rd-party dependencies of the Grapl analyzerlib:
```sh
./pants dependencies --type=3rdparty src/python/grapl_analyzerlib::
```

As the last example shows, Pants knows about and can manage the
3rd-party Python dependencies we use, across the project. The code in
the new `3rdparty` directory contains a `requirements.txt` file as
well as a `constraints.txt` file that is generated from it (see also
`build-support/generate_constraints.sh). These two files will act
analogously to a `Cargo.toml` and a `Cargo.lock` file from the Rust
world, and will be what Pants uses to satisfy dependencies. Any
existing `requirements.txt` files are still present, and will remain
so until build tasks can be migrated to Pants.

[1]: https://pantsbuild.org

Signed-off-by: Christopher Maier <chris@graplsecurity.com>