Run Pants-mediated linting in Github Actions
As the first step of the transition to Pants, we'll start linting
through Pants in CI. All previous linting logic remains in place; we
can delete it later.

As part of this, we add a `pants.ci.toml` file, containing CI-only
configuration overrides.

Additionally, we start leveraging caching in Github Actions. Pants can
cache a lot of information and artifacts, which can have a dramatic
effect on how long operations take. For now, we'll regenerate the
cache whenever our "lock file" changes. All of this can, of course, be
tweaked as we go.

Signed-off-by: Christopher Maier <chris@graplsecurity.com>