Ensure that `cargo audit` is run from the proper directory
The `actions-rs/audit-check` action doesn't appear to be able to run
in a subdirectory, and Github Actions' `working-directory` doesn't
apply to `uses` steps. Until either of those change, we'll just do it
ourselves.

Longer term, it would be nice to use that action, though, because it's
got much nicer output. In the meantime, however, this will serve our
purposes.

Signed-off-by: Christopher Maier <chris@graplsecurity.com>