Remove old py_lint.sh script
All the logic is handled by Pants now.

Additionally, our needs have moved beyond the pre-commit hook that was
present, and was the last remaining user of py_lint.sh.

(That the hook was also an incomplete solution to the problem, and
seemingly unused, are other arguments in favor of removal.)

Signed-off-by: Christopher Maier <chris@graplsecurity.com>