Run Python unit test under Pants
Here, we tune Pants to be able to run all our Python unit tests. There
are a few points to note:

First, we are still filtering by pytest markers: things that are not
marked as "integration_test" are unit tests. This was done to make as
minimal change in existing code layout and organization as
possible. It does, however, introduce a few wrinkles. Pants will look
across all the Python test files it knows about and then parse them
during the Pytest collection phase. If Pytest finds no tests that
match the markers, that is a kind of test failure. Thus, when running
tests across the entire project, we can get failures simply because a
particular project only has integration tests, for instance. This has
been addressed with the use of a Pytest
plugin (`pytest-custom_exit_code`) that allows us to treat this
situation as a passing test.

Secondly, because we are filtering on pytest markers, and because this
involves loading up Python code in order to read those markers, any
failure at this stage will also be a test failure. As it turns out,
the `test_engagement_edge.py` integration test is written in such a
way that if the `BUCKET_PREFIX` environment variable is not present,
an exception will be raised... at code loading time. Rather than
address this here, I have opted to use a Pants tag on that test (I've
called it "needs_work" for now). Because Pants tags exist at a higher
level than the source code, they can be used to restrict the inputs to
be fed to Pytest in the first place. In the new
`test-unit-python-pants` Makefile target, we actually use both this
Pants tag and Pytest marker arguments for filtering.

Because of all this, the only change I've made to actual test code is
adding `integration_test` markers where they were missing.

Longer term, I believe it would be more flexible to slightly
reorganize the tests to make it easier to target specific classes of
test (i.e. unit vs integration). This could be with specific naming
conventions on files, or separating integration tests into a separate
directory, etc. None of that is done here, though, in the interest of
keeping a light touch.

Aside from those organizational concerns, you can now run all the
existing Python unit tests _without_ Docker, and much faster by
running `make test-unit-python-pants`

Note: this does not remove or change anything in the Docker-based
system for running Python tests (yet). However, it does make it
unnecessary, since Pants runs tests in isolated environments anyway.

Signed-off-by: Christopher Maier <chris@graplsecurity.com>