Overhaul environment variable handling in Docker Compose
By sharing things in a `.env` file, we can consolidate a lot of
configuration data in a single place, reducing duplication,
eliminating drift, and making it easier to understand the state of
things.

Now, the docker-compose files capture the "structure" of our
architecture, while the `.env` file extracts specific details about
it.

Signed-off-by: Christopher Maier <chris@graplsecurity.com>