Move .env file to local-grapl.env for explicitness
In the further interest of explicitness, every command that needs this
environment to run is marked as such in the `Makefile` with a new
`WITH_LOCAL_GRAPL_ENV` code snippet that safely populates the
execution environment with the contents of this file. Please read the
documentation comments in the `Makefile` for further details.

Signed-off-by: Christopher Maier <chris@graplsecurity.com>