Ignore orphan containers in targets with multiple compose runs
When we invoke `docker-compose` multiple times with the same project
name, but different files, we see warnings about the presence of
"orphan containers". Any time `docker-compose` runs and finds services
in a project, but that aren't defined in the file(s) it is currently
operating, it regards them as "orphans".

Since we run two compose runs (once to set up an environment, and once
to execute a series of tests within that same environment), we trigger
this "orphan" detection. In our case, this is expected and not an
issue for concern.

To remove potentially confusing warning messages, we can selectively
ignore orphans in these tasks by setting task-specific environment
variable overrides.

Signed-off-by: Christopher Maier <chris@graplsecurity.com>