Consolidate localstack container usage
There's not really a reason to have separate containers for the
individual localstack services we're running. We can pull SQS and
SecretsManager back into the main localstack container to simplify
things. Additionally, localstack uses DynamoDB Local internally, so
there's no harm in dropping that as a separate container, either.

While localstack does have an S3 implementation, we are keeping MinIO
as our S3 clone. When I switched to localstack's S3, we had an
integration test failure where our UX router code claimed it couldn't find a
particular JavaScript file in the S3 bucket, even though it was
definitely there. This is "local-only" code, however, and may end up
going away as we get closer fidelity to production deployments in our
local environments. In any case, we're going to continue with MinIO
for now.

Signed-off-by: Christopher Maier <chris@graplsecurity.com>