Move all services behind Localstack's edge port
After localstack 0.11.5, all services will be behind the edge port
anyway. There's no reason we need individual ports, so this is just
preventing us from moving to more modern localstack versions.

For all the services we were sharing back out to the host, we will
continue to export the same ports we did before; we'll just use the
localstack edge port internally.

Importantly, while the documentation for localstack suggests that it
should be possible for it to act as a proxy for *external* service
replacements, I have not been able to get this to work. This should
allow us to use MinIO transparently through the localstack interface,
without having to specify a different `S3_ENDPOINT` value. This is why
we are not using `LOCALSTACK_PORT` for our S3 service now.

Signed-off-by: Christopher Maier <chris@graplsecurity.com>