Introduce Pulumi, initially for local testing infrastructure
Here, we introduce Pulumi for managing infrastructure. For now, it is
only handling a portion of infrastructure provisioning for our local
Grapl testing environment, but the long-term goal is to migrate from
CDK to Pulumi. Among other benefits, this will allow us to achieve
better parity between testing and production environments.

The strategy taken here is to encode the AWS-specific infrastructure
from `grapl_provision/grapl_provision.py` and
`grapl_provision/provision_local_identity_table.py` that we use in our
tests as Pulumi resources. This allows us to remove almost all of this
custom code (we still need to keep the Dgraph-related code until we
can create a custom Pulumi resource for this).

Some early benefits of Pulumi can be seen in
`pulumi/infra/service_queue.py`. Here, we encapsulate three separate
SQS queues into a `ServiceQueue` abstraction. The queues are linked by
their various redrive policies, and all pertain to the same Grapl
service. We can now treat the collection of queues as one logical
construct, and Pulumi will manage the dependencies between them on our
behalf.

There are a few legitimately "local-only" bits of provisioning we have
to do here; namely, the creation of our JWT secret and the creation of
a well-known Grapl user account. We can conditionally include these
resources only when operating on our `local-grapl` stack, which
ensures that these aren't included in production deployments.

To begin using this in our testing environments, we have a
`grapl-pulumi` container that performs provisioning. Our existing
`grapl-provision` container depends on this (we still need it to set
up Dgraph, remember). All other containers continue to depend on
`grapl-provision`, as before. When the functionality of
`grapl-provision` can be handled by Pulumi, we can condense all the
provisioning logic back into a single container.

Signed-off-by: Christopher Maier <chris@graplsecurity.com>