Add editor_setup.py script for generating pyrightconfig.json
Pyright is an editor integration that we can use to make it easier to
deal with our Python code (in the state that it currently exists!) in
our editors. However, setting up its configuration file for our
purposes can be complex and confusing.

All the information we need is in `pants.toml`, but that's pretty
complicated right now, and will change as time goes on. Rather than
documenting in a markdown file what works now, or including a file in
the repository that would force the same configuration on all users,
we'll use automation to help us.

Since we've now got a virtual environment for our Python dependencies,
it's not a problem to use Python to script up a solution, using the
`toml` and `click` libraries. This commit includes an
`editor_setup.py` Click application that can be used to both initially
generate a `pyrightconfig.json` file, but also to keep it updated as
the source data from `pants.toml` changes.

Signed-off-by: Christopher Maier <chris@graplsecurity.com>