Re-structure EDITOR_SETUP.md to be language-centric
As we add configuration for other editors in the future, there may be
overlap in how a given language may be configured across editors. For
example, different editors may make use of a common plugin or
configuration format (which is likely, given the rise of things like
LSP).

By keeping the documentation language centric, we can keep potentially
identical (or at least very similar) instructions close together,
which increases the likelihood that they'll be updated together.

Signed-off-by: Christopher Maier <chris@graplsecurity.com>