Fix 'make down' by running with the various project names that we use for tests (integration, e2e).
The test wrapper used to stop all containers using the same compose
project name, but this wasn't obvious and had changed. When later trying
to run 'make down', docker-compose would complain that there are still
containers using the grapl-network. This diff removes the 'stop'
funcationality into the Makefile, where we can be more explicit and
reduce the code in our home-grown wrapper.  To do this, we use .ONESHELL
in the Makefile, which normally impacts behavior of make, but we restore
that with additional shell opts.