Add a "modern-lambdas" Makefile target for use with Local Grapl
Now that Local Grapl can run lambdas natively, we need a way to ensure
that those artifacts are built appropriately when running Local
Grapl. However, we don't want to waste time building lambdas that
haven't yet been ported to Local Grapl.

To address this, we now have a `modern-lambdas` target that only
builds the lambdas that can be consumed by Local Grapl. Eventually,
this will be all lambdas, but for now it is useful to make a
distinction.

The Rust build for metric-forwarder does take a bit of time (it's Rust
code, after all), but the Docker-based build does a good job of
caching, so it shouldn't have much of an impact beyond the first
build (unless you happen to be working on the metric-forwarder, that
is). The Python lambdas are built by Pants, which does caching
natively, so the added time there should be neglible as well. As more
services run as lambdas, we'll have fewer docker builds to do and
fewer docker-compose services to spin up, so we'll be saving time
there as well.

Signed-off-by: Christopher Maier <chris@graplsecurity.com>