Add a lot of Pulumi code for lambdas, etc. running locally
Here, we get the first of our lambda-based services constructed in
Pulumi.

We have a `Lambda` abstraction, which contains the definition for an
AWS lambda and an alias, as well as functionality to optionally push
logs from that lambda to another lambda (our metric-forwarder).

Our `Lambda` is parameterized with a `LambdaArgs` abstraction, which
contains information about how to set up the function, including the
code archive, handler endpoint, resource limits, etc. We can also
subclass this to create target argument types, such as
`PythonLambdaArgs`, to allow us to centralize information like which
Python runtime we use.

Building outward, we have a `QueueDrivenLambda`, which takes an
instance of our `Lambda` and couples it with resources that allow it
to be triggered by output from an SQS queue.

This, in turn, is an important component of our `Service` abstraction,
which mirrors the same abstraction from our CDK code. It consists of a
series of SQS queues that trigger a "handler" lambda, along with a
backup "retry handler" lambda.

Finally, we come to concrete instances of our Grapl services:
`EngagementCreator` and `MetricForwarder`. These are built on top of
the previous abstractions, along with whatever associated AWS
resources are needed for their specific operation.

Currently these are set up for running in Local Grapl (notably, they
need a reference to a VPC to achieve parity with their CDK
counterparts). That will come in a follow-up PR that allows Pulumi to
read CDK stack information to gradually move real deployments over to
being fully managed by Pulumi a resource at a time.

Signed-off-by: Christopher Maier <chris@graplsecurity.com>