Expose API Gateway to the rest of Local Grapl
Our base infrastructure code doesn't define a DNS name for the API
Gateway instance. As a result, we don't have a stable address to
access that API.

To complicate things, Localstack has its own special invocation URL
for APIs.

To help abstract this, we'll put the API gateway instance behind our
existing Nginx service (if you squint, this basically turns Nginx into
a glorified DNS server for our API gateway). To do this, we have to
expose the API Gateway instance's ID to Nginx. We'll use Pulumi stack
outputs along with a Docker volume shared between the Pulumi and Nginx
containers to get the information from one place to the other
relatively straightforwardly. We write the stack output to a file in
the volume on the Pulumi side, and then read that file into an
environment variable that our Nginx configuration template uses to
create the final URL.

For now, we are only exposing the auth endpoint, as that is the only
thing the API Gateway knows about at the moment. As more lambdas are
put behind the API Gateway, our Nginx configuration will simplify
until it simply proxies everything to the gateway directly.

Signed-off-by: Christopher Maier <chris@graplsecurity.com>