Update Python dependencies for Pants 2.4.0
When we updated to Pants 2.4.0, we forgot to update our
`constraints.txt` file to account for that.

This, however, ended up pulling in an incompatible version of
setuptools.

With 2.2.0, we were pulling in setuptools 50.3.2, but with 2.4.0, we
pull in setuptools 53.1.0. However, this appears to conflict with
lambdex (the library that is the basis for the `python_awslambda`
packaging type we use), which apparently has a setuptools constraint
of <50.4,>=50.3.0 , which is unsatisfiable with setuptools==53.1.0.

I have temporarily pinned setuptools to 53.1.0 to preserve our Lambda
generation capabilities, and this seems to have no other adverse side
effects. Hopefully we can remove this pinning soon.

Signed-off-by: Christopher Maier <chris@graplsecurity.com>