Tweak makefile ordering for zip creation
Not entirely sure about this, but it looks like our caching in
Buildkite can fail on the `make zip` step because the cache directory
is still changing as the cache is being tarballed up.

By moving the pants-involved commands earlier, we can give it a chance
to "settle down" before the final caching is done.

We should keep an eye on this, however.

Signed-off-by: Christopher Maier <chris@graplsecurity.com>