Switch back to official Localstack container
Previously, we were using an internally-built Localstack container
with a patch for calling our Chalice lambdas. Since then, the real fix
has landed in the main Localstack repository (possibly addressing some
additional lambda invocation related behavior that we also need, which
was not addressed in our custom container).

For now, these fixes are only available on the `latest` containers,
but in the interest of getting back on the official codebase, we'll
use this until Localstack 0.12.11 is formally released.

Signed-off-by: Christopher Maier <chris@graplsecurity.com>