Consolidate DynamoDB read, read/write permissions
Since Roles have a maximum size of ~10Kb, and because the Node
Identifier interacts with a lot of DynamoDB tables, the approach we
had been taking of dealing with each table's permissions individually
ended up exceeding the allowed size.

Now, we consolidate all the DynamoDB permission grants into a single
policy, which greatly reduces the resulting size, while keeping the
same semantic content.

Signed-off-by: Christopher Maier <chris@graplsecurity.com>