Move grapl Pulumi project into its own directory
We'll have more projects soon, and this also allows us to leverage the
same code in our release pipeline as we use in other internal
pipelines.

At the moment, the configuration and the `__main__.py` file are the
only things in their own directory, allowing the possibility of future
projects being able to use some of the same infrastructure code. If
that proves to not be the case, the `pulumi/infra` directory could be
moved into `pulumi/grapl`.

Apart from straightforward code moves, this commit introduces a
`repository_path` function that helps to abstract the logic of
resolving other important assets relative to the repository root.

Signed-off-by: Christopher Maier <chris@graplsecurity.com>