Allow container images to be pulled from Cloudsmith
If a Fargate service has a concrete version specified in the Pulumi stack
configuration, that container image will be pulled from our Cloudsmith
repository. This will be used for, e.g., release candidate deploys to
testing and production environments.

If a version is _not_ specified, we will continue to build the image
locally, push to a newly-created ECR repository, and grant the Fargate
service permissions to read from that repository.

Note that an ECR repository is _not_ created if a version was
specified for a service.

Signed-off-by: Christopher Maier <chris@graplsecurity.com>