Build UX tarball in the merge pipeline
A new makefile target, `ux-tarball`, has been added, which creates a
`tar.gz` file containing the UX assets. This is then used in our CI
merge pipeline when we generate artifacts, pushing it to Cloudsmith
and adding the version to our Pulumi stack(s) as we do with our other
artifacts.

The Pulumi code has been modified slightly to take this into
account. One wrinkle is that Pulumi has no native way to expand a
tarball into an S3 bucket, which is how our frontend is currently
deployed. To get around this, a helper script
`pulumi/bin/prepare_grapl_ux_dependency.sh` has been added. When given
the name of a Pulumi stack, it will prepare the UX assets as
appropriate. If a version pin for the UX is present in the stack, that
version will be downloaded from Cloudsmith and expanded into
`dist/grapl-ux-<VERSION>`. If a version pin is not present, we run
`make build-ux` on the source code that is present. In either case,
Pulumi will pull the code from the correct location and deploy.

If we ever change our deployment approach, or Pulumi gets new
abilities to unpack tarballs into S3 buckets, we should *definitely*
revisit this situation.

Signed-off-by: Christopher Maier <chris@graplsecurity.com>