Temporarily disable Pants caching plugin
For as-yet-undiagnosed reasons, the caching plugin stopped working
properly, preventing Pants from being able to run. It appears to have
something to do with a recent update to the Python `requests` library.

The engineers at Toolchain Labs are aware of the issue and are working
on a fix. In the meantime, and to unblock our pipelines, we're just
going to disable the plugin. Jobs may take a bit longer to run, but
they will at least be running.

As soon as we can re-enable the plugin, we will.

Signed-off-by: Christopher Maier <chris@graplsecurity.com>