Switch over to new packer-staging queue for Packer builds
Rather than building AMIs in our CI/CD account, we'll build them in
our new Staging account.

We need to pass AWS auth variables because of how the build nodes are
managing their authentication; future changes will simplify this, but
for now, this is what we need to do.

This also will allow us to retire our old "packer" queue.

Signed-off-by: Christopher Maier <chris@graplsecurity.com>