Drop IAM role after building AMIs, before uploading artifacts
This is a stop-gap measure to ensure that our manifest files are going
to the proper place (our private artifact bucket). It is
emphatically *not* intended to be a permanent solution.

While we run these Packer jobs in Docker containers, and until a bug
is fixed in the Buildkite agent (no issue to track yet, but it has
been reported and confirmed through Buildkite Support), this is the
most straightforward solution.

Read documentation comments for further background on this, as well as
https://github.com/grapl-security/issue-tracker/issues/620

Signed-off-by: Christopher Maier <chris@graplsecurity.com>