Add provisioning pipeline
Deploys the `grapl/grapl/testing` Pulumi stack to our Staging AWS
account.

As currently structured, this will update the existing testing
infrastructure with whatever changes are coming through the pipeline;
it does not reprovision the environment from scratch. (This is
something we may wish to add in the future, however.)

Includes a mock testing pipeline as well; eventually this will be for
running tests against that infrastructure.

Signed-off-by: Christopher Maier <chris@graplsecurity.com>