Change name of managed virtualenv to simply "venv"
Now that we're using shared scripts to manage our Pulumi operations in
CI, it's even more important that we adhere to the common conventions
of those scripts.

Here, that means using the same name for the Python virtual
environment in our `manage_virtualenv.sh` script. It doesn't need to
be "grapl-venv" anymore (you can probably assume it's for Grapl since
it's in the `grapl` repository, after all).

You'll want to run `build-support/manage_virtualenv.sh populate` after
this lands, and activate the new one. As always, following the
instructions output by the script will help you out with this.

Signed-off-by: Christopher Maier <chris@graplsecurity.com>