osquery-subgraph-generator => osquery-generator
This was initially motivated because our Pulumi code assumed an
artifact named "osquery-generator", but our container name (and all the
code underlying it) was named "osquery-subgraph-generator".

To ensure a consistent mapping between artifact and service in our
Pulumi code, and an overall consistency in the codebase in general, we
have removed the "subgraph" part of the name.

Signed-off-by: Christopher Maier <chris@graplsecurity.com>