Stop container builds of graphql-endpoint and model-plugin-deployer
These two services are the only Lambda ZIP-deployed services that we
are not actually running as Lambdas in Local Grapl; for that
environment, these run as separate container services.

Unfortunately we were erroneously building these containers in CI and
uploading them to our artifact repository. As a result, the version
pins for these Lambda services in our Pulumi stack configurations were
frequently getting masked by the new container versions (we're still
building the Lambda versions in CI, but only when their code changes;
we don't have as good tooling to do that with the containers yet, so
they get built every time through the pipeline.)

As a result of this, Pulumi updates will fail because Cloudsmith does
not have a "raw" artifact for these service/version pairs. It does
have a container artifact for them, but these are two different
download paths (you don't download a ZIP file using `docker pull`, and
you don't pull a container using `curl`).

All we do here is remove these containers from the list of services we
build containers for in CI, and add some more comments so this doesn't
trip us up again before we can remove those containers altogether.

Signed-off-by: Christopher Maier <chris@graplsecurity.com>