Pass the Notebook name to engagement-edge directly
Rather than generating the name internally, `engagement-edge` will now
simply receive the notebook instance name via
`GRAPL_NOTEBOOK_INSTANCE` in the environment.

We do need to keep around a tiny bit of logic to account for the fact
that Localstack doesn't support sagemaker, but it's minor, in the
grand scheme of things.

Signed-off-by: Christopher Maier <chris@graplsecurity.com>