Use the debug version of our distroless containers for Rust services
If we're going to be able to pass in values from Pulumi outputs in our
Local Grapl Rust services, we need to have minimal access to a shell
and `cat`. Using the `debug` variant of the distroless container
basically just adds Busybox, which is all we need.

It's unfortunate that this is only needed for Local Grapl because
we're not able to run our Fargate services in Localstack. Once we
switch over to Nomad, this won't be an issue. In the meantime, adding
Busybox is a minimal concession to make.

Signed-off-by: Christopher Maier <chris@graplsecurity.com>